requirements.txt
# os
streamlit
openai
python-dotenv